+++ pytorch_lightning/trainer/distrib_data_parallel.py
@@ -130,23 +130,17 @@ When the script starts again, Lightning will:
 import os
 import re
 from abc import ABC, abstractmethod
-from distutils.version import LooseVersion
-from typing import Union, List, Optional, Callable, Tuple
-import subprocess
-import sys
-from time import sleep
-import numpy as np
-from os.path import abspath
-from pkg_resources import parse_version
+from typing import Union, List, Optional, Tuple
 
 import torch
+
 from pytorch_lightning import _logger as log
-from pytorch_lightning.loggers import LightningLoggerBase
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.distributed import rank_zero_warn, rank_zero_info
 from pytorch_lightning.core.datamodule import LightningDataModule
 from pytorch_lightning.core.lightning import LightningModule
-
+from pytorch_lightning.loggers import LightningLoggerBase
+from pytorch_lightning.utilities.cloud_io import atomic_save
+from pytorch_lightning.utilities.distributed import rank_zero_warn, rank_zero_info
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 try:
     from apex import amp
@@ -167,10 +161,6 @@ except ImportError:
     XLA_AVAILABLE = False
 else:
     XLA_AVAILABLE = True
-
-PID = os.getpid()
-RNG1 = np.random.RandomState(PID)
-RANDOM_PORTS = RNG1.randint(10000, 19999, 1000)
 
 
 class TrainerDDPMixin(ABC):
@@ -243,9 +233,6 @@ class TrainerDDPMixin(ABC):
         """Warning: this is just empty shell for code implemented in other class."""
 
     def init_tpu(self):
-        # turn off all the GPU stuff
-        # self.distributed_backend = 'tpu'
-
         # enable tpu
         self.use_tpu = True
 
@@ -397,22 +384,6 @@ class TrainerDDPMixin(ABC):
         # don't make this debug... this is good UX
         rank_zero_info(f'CUDA_VISIBLE_DEVICES: [{os.environ["CUDA_VISIBLE_DEVICES"]}]')
 
-    def set_random_port(self, force=False):
-        """
-        When running DDP NOT managed by SLURM, the ports might collide
-        """
-        # pick a random port first
-        assert self.num_nodes == 1, 'random port can only be called from single node training'
-        global RANDOM_PORTS
-        default_port = RANDOM_PORTS[-1]
-        RANDOM_PORTS = RANDOM_PORTS[:-1]
-
-        # when not forced, use the user port
-        if not force:
-            default_port = os.environ.get('MASTER_PORT', default_port)
-
-        os.environ['MASTER_PORT'] = str(default_port)
-
     def transfer_distrib_spawn_state_on_fit_end(self, model, mp_queue, results):
         if self.distributed_backend.lower() not in ['ddp_spawn', 'ddp_cpu', 'tpu']:
             return
@@ -432,13 +403,7 @@ class TrainerDDPMixin(ABC):
             last_path = None
             if not self.testing and best_model_path is not None and len(best_model_path) > 0:
                 last_path = re.sub('.ckpt', '.tmp_end.ckpt', best_model_path)
-                # Can't use the new zipfile serialization for 1.6.0 because there's a bug in
-                # torch.hub.load_state_dict_from_url() that prevents it from loading the new files.
-                # More details can be found here: https://github.com/pytorch/pytorch/issues/42239
-                if LooseVersion(torch.__version__).version[:3] == [1, 6, 0]:
-                    torch.save(model.state_dict(), last_path, _use_new_zipfile_serialization=False)
-                else:
-                    torch.save(model.state_dict(), last_path)
+                atomic_save(model.state_dict(), last_path)
             mp_queue.put(last_path)
 
     def save_spawn_weights(self, model):