+++ tests/models/test_amp.py
@@ -8,13 +8,16 @@ import wandb
 import tests.base.develop_pipelines as tpipes
 import tests.base.develop_utils as tutils
 from pytorch_lightning import Trainer
+from pytorch_lightning.trainer.states import TrainerState
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
 
 
 @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
 def test_multi_gpu_wandb_ddp_spawn(tmpdir):
-    """Make sure DP/DDP + AMP work."""
+    """
+    Test ddp + wb
+    """
     from pytorch_lightning.loggers import WandbLogger
     tutils.set_random_master_port()
 
@@ -173,3 +176,45 @@ def test_cpu_model_with_amp(tmpdir):
 
     with pytest.raises((MisconfigurationException, ModuleNotFoundError)):
         tpipes.run_model_test(trainer_options, model, on_gpu=False)
+
+
+def test_amp_without_apex(tmpdir):
+    """Check that even with apex amp type without requesting percussion=16 the amp backend is void."""
+    os.environ['PL_DEV_DEBUG'] = '1'
+    model = EvalModelTemplate()
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        amp_backend='native',
+    )
+    assert trainer.amp_backend is None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=1,
+        amp_backend='apex',
+    )
+    assert trainer.amp_backend is None
+    trainer.fit(model)
+    assert trainer.state == TrainerState.FINISHED
+    assert trainer.dev_debugger.count_events('AMP') == 0
+
+
+@pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires GPU machine")
+def test_amp_with_apex(tmpdir):
+    """Check calling apex scaling in training."""
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    model = EvalModelTemplate()
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=1,
+        precision=16,
+        amp_backend='apex',
+        gpus=1,
+    )
+    assert str(trainer.amp_backend) == "AMPType.APEX"
+    trainer.fit(model)
+    assert trainer.state == TrainerState.FINISHED
+    assert trainer.dev_debugger.count_events('AMP') == 10