+++ tests/trainer/test_trainer_steps_result_return.py
@@ -7,6 +7,7 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
+from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
 from pytorch_lightning.core.step_result import TrainResult
 from tests.base import EvalModelTemplate
 from tests.base.deterministic_model import DeterministicModel
@@ -397,7 +398,7 @@ def test_no_auto_callbacks_with_train_loop_only(tmpdir):
     )
     trainer.fit(model)
 
-    all_losses = trainer.dev_debugger.saved_losses
+    all_losses = trainer.dev_debugger.saved_train_losses
     assert len(all_losses) == batches * epochs
 
     assert trainer.checkpoint_callback.monitor == 'checkpoint_on'
@@ -438,7 +439,7 @@ def test_no_callbacks_with_train_loop_only(tmpdir):
     )
     trainer.fit(model)
 
-    all_losses = trainer.dev_debugger.saved_losses
+    all_losses = trainer.dev_debugger.saved_train_losses
     assert len(all_losses) == batches * epochs
 
     assert trainer.early_stop_callback is None
@@ -477,7 +478,7 @@ def test_use_callbacks_with_train_loop_only(tmpdir):
     assert len(early_stop_vals) == num_expected_epochs
     min_val = min([x['best'] for x in early_stop_vals])
     assert min_val == 171 + 9
-    all_losses = trainer.dev_debugger.saved_losses
+    all_losses = trainer.dev_debugger.saved_train_losses
 
     from collections import Counter
     batch_idxs = Counter([x['batch_idx'] for x in all_losses])
@@ -534,6 +535,41 @@ def test_full_train_loop_with_results_obj_dp(tmpdir):
     assert 'epoch_train_epoch_end_metric' in seen_keys
 
 
+@pytest.mark.skipif(torch.cuda.device_count() < 2, reason="test requires multi-GPU machine")
+def test_loop_steps_only_dp(tmpdir):
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    batches = 10
+    epochs = 3
+
+    model = EvalModelTemplate()
+    model.validation_step = None
+    model.test_step = None
+    model.training_step = model.training_step_result_obj_dp
+    model.training_step_end = None
+    model.training_epoch_end = None
+    model.validation_step = model.validation_step_result_obj_dp
+    model.validation_step_end = None
+    model.validation_epoch_end = None
+    model.test_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        distributed_backend='dp',
+        gpus=[0, 1],
+        max_epochs=epochs,
+        early_stop_callback=True,
+        row_log_interval=2,
+        limit_train_batches=batches,
+        weights_summary=None,
+    )
+
+    trainer.fit(model)
+
+    assert model.training_step_called
+    assert model.validation_step_called
+
+
 def test_result_map(tmpdir):
     result = TrainResult()
     result.log_dict({'x1': torch.tensor(1), 'x2': torch.tensor(2)})
@@ -543,3 +579,43 @@ def test_result_map(tmpdir):
     assert 'x2' not in result
     assert 'y1' in result
     assert 'y2' in result
+
+
+def test_result_monitor_warnings(tmpdir):
+    """
+    Tests that we warn when the monitor key is changed and we use Results obj
+    """
+    model = EvalModelTemplate()
+    model.test_step = None
+    model.training_step = model.training_step_result_obj
+    model.training_step_end = None
+    model.training_epoch_end = None
+    model.validation_step = model.validation_step_result_obj
+    model.validation_step_end = None
+    model.validation_epoch_end = None
+    model.test_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=2,
+        early_stop_callback=True,
+        row_log_interval=2,
+        limit_train_batches=2,
+        weights_summary=None,
+        checkpoint_callback=ModelCheckpoint(monitor='not_val_loss')
+    )
+
+    with pytest.warns(UserWarning, match='key of ModelCheckpoint has no effect'):
+        trainer.fit(model)
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=2,
+        row_log_interval=2,
+        limit_train_batches=2,
+        weights_summary=None,
+        early_stop_callback=EarlyStopping(monitor='not_val_loss')
+    )
+
+    with pytest.warns(UserWarning, match='key of EarlyStopping has no effec'):
+        trainer.fit(model)